refactor(sys): always use integer types from arch/types.h
Use those instead of primitive types such as short, int or long.

+++ sys/include/console.h
@@ -7,22 +7,22 @@
 
 /* Prints a message to serial console and returns the amount of printed
    characters. */
-unsigned int consolewrite(char *m);
+u16 consolewrite(char *m);
 
 /* Prints an integer to serial console in binary form and returns the amount of
    printed characters. */
-unsigned int consolewriteintb2(u64 n);
+u16 consolewriteintb2(u64 n);
 
 /* Prints an integer to serial console in signed decimal form and returns the
    amount of printed characters. */
-unsigned int consolewriteintbase10s(s64 n);
+u16 consolewriteintbase10s(s64 n);
 
 /* Prints an integer to serial console in unsigned decimal form and returns the
    amount of printed characters. */
-unsigned int consolewriteintb10u(u64 n);
+u16 consolewriteintb10u(u64 n);
 
 /* Prints an integer to serial console in hexadecimal form and returns the
    amount of printed characters. */
-unsigned int consolewriteintb16(u64 n);
+u16 consolewriteintb16(u64 n);
 
 #endif
